further modularization; added gitignore
